Add storage

+++ start_app_from_scratch.py
@@ -1,29 +1,42 @@
 import logging
 import os
-import shutil
 import subprocess
 import sys
 
-from src.package_wielkate.main.commons.constants import DATABASE_NAME, IMAGES_DIRECTORY
+import requests
+
+from src.package_wielkate.main.commons.constants import DATABASE_NAME, CLOTHES_MATCHING_API
 
 program1 = 'src/package_wielkate/main/scripts/prepare_colors.py'
 program2 = 'src/package_wielkate/main/scripts/prepare_combinations.py'
 program3 = 'src/package_wielkate/main/__main__.py'
 python = sys.executable
 
+GREY_LIGHT = '\033[37m'
+WHITE = '\033[97m'
+GREEN = '\033[92m'
+CYAN = '\033[36m'
+
 logging.basicConfig(
     level=logging.INFO,
     datefmt='%Y-%m-%d %H:%M:%S',
-    format="%(asctime)s  %(name)s  [%(levelname)s]: %(message)s"
+    format=f"{CYAN}%(asctime)s  {GREY_LIGHT}%(name)s  {GREEN}[%(levelname)s]: {WHITE}%(message)s"
 )
 logger = logging.getLogger(__name__)
 
 
+def clear_bucket():
+    requests.delete(f'{CLOTHES_MATCHING_API}/delete')
+
+
 def run_programs():
-    os.remove(DATABASE_NAME)
-    logger.info('Remove database')
-    shutil.rmtree(IMAGES_DIRECTORY)
-    logger.info('Remove image directory')
+    if os.path.exists(DATABASE_NAME):
+        os.remove(DATABASE_NAME)
+        logger.info('Remove database')
+    else:
+        logger.info('Database file not found, skipping removal')
+    clear_bucket()
+    logger.info('Delete all images from the bucket')
     subprocess.run([python, program1])
     logger.info('Prepare colors table')
     subprocess.run([python, program2])
